Move biz/handler/api/ws to /ws
